Cleanup, first pass.
This cleanup is a significant refactoring of the original
ttfunk codebase. It removes the dependence on method_missing,
and adds explicit attribute lists. It also avoids dynamic
require statements, preferring to explicitly list the files
to load.

This cleanup also adds support for the 'loca', 'post', 'OS/2',
and 'glyf' TrueType tables.