Fix JRuby --1.9 problem with subsetting of non-low-ASCII chars
See the failing spec in Prawn:

https://github.com/prawnpdf/prawn/commit/386998effa7ce299c749f47063d7ff426b8f980a

When a string to be encoded in a TTF font contained low-ASCII as well as
non-low-ASCII characters that could both be in the same subset, we would
be at the mercy of whatever encoding `char.chr` decided to return for
the first character of the segment. If, for example, this was US-ASCII,
and a subsequent character in the subset was not US-ASCII, we would
receive an encoding error when trying to shovel the new character as a
Fixnum (>= 128) into the US-ASCII string.

Resolve the problem by forcing the encoding on the string to ASCII-8BIT
(BINARY), so that Ruby treats the subset strings as a stream of bytes
without imputing encoding semantics.