Fix loca table serialization
The loca table serialiation code didn't properly detect the suitable
table format which led to corrupted subset font.

The issue was that the code didn't detect misaligned glyphs properly.
When a glyph's offset is not dividable by 2 short format offset led to
impropr resulting offset. This prevented some glyphs from being found in
the font.

Tangebly, the code improperly checked if the actual affset fit into
16-bit field instead of short offset, which didn't result in corrupt
loca table but could make it bigger than it actually neded to be.