Fix maxp table
Original approach was right but there was an error in field sizes. This
lead to shift of the following field. One of the fields is
maxStackElements which defines stack size for font program. FreeFont
(also used in Ghostscript) is very strict about stack size. Incorrect
maxStackElements value lead to stack overflow in FreeFont. Consequently,
Ghostscript couldn't parse the font and completely discarded it.