Add long date time encoding/decoding helpers
The created and modified values in the TTF header is a custom type
called a long date time. A definition can be found here:
https://developer.apple.com/fonts/TrueType-Reference-Manual/RM06/Chap6.html

The bytes are big endian encoded, and the basis is offset from standard
Epoch. To make these values at all useful in Ruby, they need to be
properly decoded from bytes and then offset corrected to get a Time.
Similarly, we need to do the reverse transition when encoding.